Eliminated the faults noted in the remarks
