+++ styles/widgets/generic/treeView.generic.less
@@ -68,6 +68,9 @@
     @GENERIC_TREEVIEW_ARROW_ICON_SIZE: 18px;
 }
 
+@GENERIC_TREEVIEW_SEARCH_EDITOR_MARGIN: 6px;
+@GENERIC_TREEVIEW_SEARCH_EDITOR_HEIGHT: round(@GENERIC_BASE_LINE_HEIGHT * @GENERIC_BASE_FONT_SIZE) + @GENERIC_BASE_INLINE_BORDEREDWIDGET_TOP_PADDING + @GENERIC_BASE_INLINE_BORDEREDWIDGET_BOTTOM_PADDING + 2*@GENERIC_BASE_BORDER_WIDTH;
+
 .dx-treeview-node-loadindicator {
     top: @GENERIC_TREEVIEW_LOADINCICATOR_OFFSET_TOP;
     left: @GENERIC_TREEVIEW_LOADINCICATOR_OFFSET_LEFT;
@@ -75,11 +78,15 @@
     height: @GENERIC_TREEVIEW_LOADINCICATOR_SIZE;
 }
 
-.dx-treeview-search {
-    margin-bottom: 6px;
-}
-
 .dx-treeview {
+    .dx-treeview-search {
+        margin-bottom: @GENERIC_TREEVIEW_SEARCH_EDITOR_MARGIN;
+    }
+
+    &.dx-treeview-with-search > .dx-scrollable {
+        height: calc(~"100% - "@GENERIC_TREEVIEW_SEARCH_EDITOR_HEIGHT + @GENERIC_TREEVIEW_SEARCH_EDITOR_MARGIN);
+    }
+
     &.dx-treeview-border-visible {
         border: 1px solid @treeview-border-color;
